requirements.txt
numpy
imgaug
keras
pillow
opencv-python
tensorflow